adds separate rule for plotting results
